Create GeneratorV2
This is sufficiently different enough in its execution of the concept that I felt it deserved to be in a different file. This version allows printing each monkey's string to be optional and theoretically supports any character (although having something like a newline character in the string being passed to main() allows the monkeys to guess characters like that, which makes the output look stupid at times).